Add DeviceState.isReady
Allow plugins to indicate when a device is fully ready for use via the
isReady flag in DeviceState. This avoids the need to check the nested
flow of DeviceInfo on ConnectedDevice, and can also encompass other ways
of being ready, such as being fully booted.

Add the bootStatusFlow to facilitate tracking boot status of devices.

Bug: 317411917
Test: n/a
Change-Id: I0067bcf00c965ce0c45007582c4a7d1c970a6b14