Limit session samplers to applicable tasks
Not every task uses the data collected from the session
samplers. As of this change, only the Live View and
Java/Kotlin Allocations tasks do. Thus, the call to start
the session samplers is now only made if (1) the task based
ux feature is turned off or (2) task based ux is on and the
task type of the initiated session is of a task that uses
the session sampler data.

Fixes: 319137535
Test: manual
Change-Id: I127eb4f1bd1c9b2ed2756f5008db78e6ab107029