318986919: Don't hang on to file contents in quickfix machinery
This CL removes the `initialText` field from the `PendingEditFile` data
structure which was found in field leak reports. On the IDE side, we
don't need it anyway -- it's unused there since we apply the deltas
directly to the (possibly edited) source buffer.

Instead, we're now passing around a separate state object which caches
file contents; this is used only during fix construction (and fix
application on the CLI side which is done immediately).

Fixes: 318986919
Test: Existing
Change-Id: Idc351363209b2bab4df4f50347a3ab496ae08d0d