Make lint warn multiple times for multiple copies of an unused resource
Previously, if lib1 had an unused resource and lib2 had the same unused
resource, running lint from a downstream app with checkDependencies true
would only result in a single warning.

This fix only fixes the case when running lint with partial analysis.
The added test in UnusedResourceDetectorTest will break if
.testModes(TestMode.PARTIAL) is deleted.

Bug: 200577800
Test: UnusedResourceDetectorTest
Change-Id: I53713d771bcb4b90166b885fdc91820f6e60a98b