Add lint jars to AndroidProject model for module libraries
This allows the artifactView call to be external scoped for lint
jars, which should make the overall process faster.

Lint jar is now unset for the project library model. For backwards
compatibility the IDE will look up the project model first, and then
the dependency model.

Also this uses the lintPublish configuration directly instead of the
published artifact as a further optimization, so it's expected some
snapshots will change and not have the lint jar anymore when it's not
relevant.

Test: Snapshot tests
Bug: 320435751
Change-Id: I5b9ba004bb9313d3cfbd112d14f2239d67dc0a38