Run device reads on device scope
Currently, Resolution.readFromDevice() has the potential to hang if it
is not run from the scope of the ConnectedDevice: if the Flow is
returned before the connection breaks, then readFromDevice() will
attempt to collect from it until its scope is cancelled.

Thus, move all I/O off the LocalEmulatorDeviceHandle's actor coroutine
to avoid blocking it, and run it in the device scope instead.

Bug: 320514147
Test: LocalEmulatorProvisionerFactoryTest, 1000 runs
Change-Id: I85393494569d2ff455df92e25325c808acb1bda9