Lint: report every resource location (part 2)
UnusedResourceDetector was previously fixed to report an incident for
each declaration location (when a resource had duplicate declarations
across modules) for partial analysis. This change further fixes
UnusedResourceDetector for global analysis mode. The relevant function
was also somewhat reorganized (with minor changes) so that the fix was
easier to make.

Bug: 200577800
Test: existing updated
Change-Id: I9ee72c570e8d7fc230ef64c11861829f5d981b12