Use default action's intent  category when launching main activity
When launching the main activity through adb shell, we need to pass the intent
action and categories as arguments. We are making a change in the logic
to find the intent category to use for launching main activity. Instead of
choosing the first category of first intent-filter declared in manifest,
we now specifically choose the intent-filter with the default action and
then pick the first category.

Bug: 315855033
Test: ActivityTest
Change-Id: Ide5fd323dd5ce9e02edd74939735c3630af3aac6