adblib: Simplify JdwpPacketReceiver implementation
The "activation" part of JdwpReceiver has always been complex, and
prone to timing issues (e.g. change id
I96574e1998634066c923e399612938858bd19715).

This commit addresses this with the following changes:

* Introduce a new `onSubscription` API, re-using the underlying
`StateFlow.onSubscription` API. This new API enables removing
the (complex) activation block logic from `JdwpPacketReceiver`,
as there is no need to send "fake" `JdwpPacket` to the shared flow
anymore, as we can simply use the `onSubscription` action to
start the `activation` block.

* Remove the "replay cache" property from `SerializedSharedFlow`, as
it was never used.

Note: This change was tested with the following commands to ensure
no flakiness is introduced:

    bazel test //tools/base/adblib-tools:adblib.tools.tests --runs_per_test=5000
    bazel test //tools/base/adblib-ddmlibcompatibility:adblib.ddmlibcompatibility.tests --runs_per_test=500

Bug: n/a
Test: Included + Existing
Change-Id: Iea50cc503a3d8d45ae04530c2822b619c259e79d