Improve log messages in `JdwpProcessPropertiesCollector`
Some `JdwpProcessPropertiesCollector` log messages always list
a process name as `null` since the process name is not yet known.
Make it clear that process name imformation was simply not
received yet. Also include info on whether WAIT command
was received. Note that the pid of the process is logged with every
`JdwpProcessPropertiesCollector` log message.

Test: n/a
Bug: 323825558
Change-Id: Iaf19e1390b096169dfa04ba93d6bd9a94408f57c