Use Application Classloader
Using Looper.getMainLooper().getThread().getContextClassLoader()
doesn't work with apps that use `android:process`. We get this in Logcat:

```
  ClassLoader.getResourceAsStream: The class loader returned by Thread.getContextClassLoader() may fail
  for processes that host multiple applications. You should explicitly specify a context class loader.
```

Using the classloader associated with the Application object seems to
work.

We don't have access to an Android Context in the inspector stack, so we
use `ArtTooling` to get a hold of the Application object and use its
classloader.

Fixes: 324033365
Test: NA
Change-Id: I9570ffb141eb6e08d41b486639474c2155634d5c