[SDK Index] Do not generate link when url is blank
Is it possible that the SDK Index xnapshot will start containing issues
for libraries that are not registered in the SDK Index. For those
libraries, the indexUrl link will be a blank string.

Android Studio currently only checks for null values and assumes that
anything else is a valid address and thus generates a link when a
library has issues from the SDK Index. After this change, there will be
a link only if the result of indexUrl is not null and not blank.

Bug: 324946212
Test: Updated GooglePlaySdkIndexTest
Change-Id: Ie1e7b47bed823173670c22a47be4337ac3d1969a