adblib: Improve thread-safety
The `ReferenceCountedFactory` class is used to create new
instances of `SharedJdwpSessionImpl` on-demand, and
close them when not used anymore.

This change improves Thread-Safety of this class by ensuring
calls to `retain()` and `release()` are serialized and never
overlap, so that it is not possible to try to open a new session
while the previous one is still being released.

The implementation also changes the way `close` is implemented
so that `close` ensure any pending (and future) operation is
"promptly" cancelled.

Bug: 324474436
Test: Included
Change-Id: I30f98c677e3ece194b713c5117b3d1ae19f0533c