Support If-Modified-Since (and similar mechanisms) in NetworkCache
If the cache timestamp is older than the cache expiry time, we open
the URL connection, but do not necessarily receive any data if the
resource is not modified since the timestamp of our cached
version (and the server at the other end of the connection supports
sending a suitable response).  If the resource was not modified,
update the timestamp of the cached file.

Update implementations of NetworkCache in this repository to extend
the new method rather than the old.  Preserve the old method for other
implementations to extend instead.

Bug: 328020192
Test: added, manually single-stepped through
Change-Id: Ia4a9d9e6e8714cbc6ddb79c0eebabfc7cfcd26d6