Create a Dedicated Project For App Inspectors
The App Inspector agents are small part of the /tools/base IntelliJ project. Since /tools/base is very large, it makes managing the code difficult.

This cl creates a small isolated project just for the inspector agents. It contains only new project level files with no changes to the actual modules. The libraries names were preserved in order to not make any changes to the existing modules.

Also note that this has no effect on Bazel support. It's only for accessing code in IntelliJ

Bug: NA
Test: NA
Change-Id: I54356845f8612c2f4d06348a45a6111374d417da