Add The Rest Of The Database Inspector Tests
All the rest of the tests from the original androidx project are included here.

The tests all passed without any changes except for a few tests in `InvalidationTest` which required calls to `shadowLooper.runToEndOfTasks()` in order to consume the next event.

This in turn required a slight change to  `SqliteInspectorTestEnvironment`

There are a few tests that were marked as flaky and were ignored.

1. `test_throttling` is indeed flaky. Needs more investigation.
2. `test_exec_hook_methods` is flaky but only because it runs 3 tests in succession and runs into the same throttling issue that causes the flakiness in `test_throttling`. I split it into 3 tests and it runs fine.

Bug: NA
Test: NA
Change-Id: I1dca3a5d7f96e021929bac2a32d0c1af3c67cf3c