Remove two tests that became coverage incompatible
ag/26502132 and ag/26502385 converted network and backgroundtask
inspector tests to Roboelectric respectively, but this means those tests
no longer produce coverage data. Since they were still running as part
of the coverage suite, this caused post-processing to fail as the
expected data file was not present. This CL removes the tests from the
coverage suite.

Bug: N/A
Test: verified locally
Change-Id: I91b21ff6fe3fcc1529d345bb600e96588ecbf8e5