Better error message for parse failures in bazel parser
Include information about which file the parse error is in, if it is
in a file.  This matters because the parser is incapable of parsing
triple-quoted strings (as used in e.g. module docstrings), and if a
user forgets to create an empty BUILD file in their new module under
tools/adt/idea or tools/base before running iml_to_build, we walk up
the directory tree and find tools/BUILD.bazel which contains such a
module docstring.

Bug: none filed
Test: existing
Change-Id: I93eb4079f8c5098ae0d7eb13b7a8d53551d0310a