adblib: fix bug with orphan JDWP process
The `JdwpProcessFactory` would sometimes create
`JdwpProcessImpl` entries for JDWP processes that
have already terminated on the the device, leading to
failure to open JDWP session every 2 seconds for
each "orphan" process (until the device is disconnected)

This started to happen when the concept of "delegate
adb session" was introduced to allow sharing a single
JDWP connection to a given process between multiple
`AdbSession` instances.

This change addresses the issue by making sure
each `JdwpProcessFactory` keeps track of the list
of active JDWP process IDs, and uses that list to
cancel `JdwpProcess` instances that should not
be active anymore.

Fixes: 322467516
Test: Included
Change-Id: Ic1a7172ec9973f4f47498672213e2e5ba08ed30b