Don't use ValueSource as transform input
If a transform input is a Property<String> computed from a
FileCollection through a ValueSource as the intermediate step, then
the build may deadlock when the inputs of two instances of that
transform are being resolved (see
https://github.com/gradle/gradle/issues/28566 for details).

This bug currently impacts AGP 8.3+ only, possibly because AGP 8.3 has a
number of cleanups / refactorings which create the conditions for this
bug to show up.

To avoid this bug, we have 2 options:
  1. Keep the transform input's type as Property<String>, but compute
     its value directly from the FileCollection without going through an
     intermediate ValueSource.
  2. Set the FileCollection as the transform input, and compute the
     String value from it inside the transform action.

Option 1 is simpler, but it will add the FileCollection to configuration
cache inputs. Therefore, we will go with option 2 in this commit.

The transform input that directly causes this bug is
`BaseDexingTransform.libConfiguration`. We will remove ValueSource from
that input and similar transform inputs which are computed from the
`coreLibraryDesugaring` configuration.

Note that there are still instances of ValueSource being used as inputs
of tasks and other transforms. We're not removing them as we can't
confirm if they will have the same issue or not, and having ValueSource
is still useful.

Bug: 329346760
Test: Added a test in CoreLibraryDesugarTest
Change-Id: I98ac834d1a21c64ce20307cd91eb4b6552e5f6ea