Propagate required SDKs to consumers
When testing locally, required SDKs are propagated to apiElements and
runtimeElements of the privacy sandbox SDK, meaning that consuming apps
don't need to redeclare those dependencies.

Also add validation to the build that only asars can be added to
requiredSdk and optionalSdk elements, to avoid propagating dependencies
that would be packaged in the app.

Bug: 265306079
Test: PrivacySandboxMediatorSdkTest
Change-Id: I787033ded8eface6c351ad92d199b304338b4cea