adblib: fix bug with orphan JDWP process
The `JdwpProcessFactory` would sometimes create
`JdwpProcessImpl` entries for JDWP processes that
have already terminated on the the device, leading to
failure to open JDWP session every 2 seconds for
each "orphan" process (until the device is disconnected)

This started to happen when the concept of "delegate
adb session" was introduced to allow sharing a single
JDWP connection to a given process between multiple
`AdbSession` instances.

This change addresses the issue by making sure
each `JdwpProcessFactory` keeps track of the list
of active JDWP process IDs, and uses that list to
cancel `JdwpProcess` instances that should not
be active anymore.

Note: This is commit is the same as
commit a79dea0a757c398722c36895214d7cfa083ad6f0 that was reverted in
commit 2d2bf98bc1def5f7c611e7c42cbef4997c71906b, but with a few
additional fixes to address flakiness exposed by tests.

Fixes: 322467516
Test: Included
Change-Id: Ibf5ea08f130d542c61b376e1ee7d530382fa9b32