Add is_read_only to DatabaseOpenedEvent
On Android, when a database is opened as Read-Only, it is not necessarily actually read-only. It is just not guaranteed to be writable.

If we do end up with an actual read-only database, we should be able to show the user
that the db was opened as read-only.

This change adds a `is_read_only` field to the DatabaseOpenedEvent, and it also changes logic that sends this event:

Whenever a db is opened or closed, we check if the type (forced, rea-only or writeable) has changed since the last time we sent an event for this DB.

Bug: NA
Test: Added
Change-Id: I436c1d2b18667063ed25d400ab75371069d03f1c