Add java and kotlin gen sources to anchor task.
GenFolderApiTest will now tests functionalities using both the
old and the new variant APIs. There is now 2 build files in the project
old_variant_api.build.gradle for testing with the old APIs and
new_variant_api.build.gradle for testing with the new variant APIs.

The test uncovered an issue with the new Variant API where we were
not registering the java and kotlin generated sources providers as
dependents of the genSource anchor task.

Test: integration tests.
Bug: N/A
Change-Id: I0bc33555a411ef42f82b4580f3e1e9d70708c924