ApkOutputProvider API to expose Apks for local deployment
This change introduces ApkOutputProvider API to provide an
ordered collection of Set of APKs that are part of Application
installation on a device. There are a number of APKs besides
Application's main APK that must be installed on a device when an
application depends on privacy sandbox runtime enabled SDKs.
ApkOutputProvider is available through ApplicationVariant
methods that configures a custom task property and sets up task
inputs correctly. InstallVariantTask and InstallVariantViaBundleTask are refactored to use this new API.

Bug: 296837048
Test: DefaultApkProviderIntegrationTest, InstallVariantTaskTest
Change-Id: I41aa3ac1bfaeaf6bce65cf0850a4bbcb6f8fc700