[SDK Index] Do not show URL for non SDK Index libraries
The provided link is to prevent older versions of Studio to get an empty
link and thus there is no point in generating a link if it does not
provide information. With this change first we first check if the link
does not come from the SDK Index and and otherwise returns it.

Bug: 339094763
Test: GooglePlaySdkIndexTest
Change-Id: I9b07c9a1ab4d7bb370572f3cd16bb9bf9a76046b