Remove name from AvdInfo constructor
The "name" of an AVD is simply the basename of the AVD's ini file; it is
already contained within the iniFile parameter. Requiring it as a
distinct parameter is extra work and only adds an opportunity for error.

Note that most code should not be using the name anyway. For display to
the user, the display name should be used instead; for a stable
identifier (that stays the same across renames), getId() should be used
instead. The AVD name is still used to identify the device on the
command line.

Bug: 339875083
Test: existing
Change-Id: I3d1c3b2783cb93066a773adecbb11cfaad40deb3