Introduce AndroidGlobalTask to pair with AndroidVariantTask
The current task hierarchy in AGP is as follows:
     BaseTask
     |--- AndroidVariantTask: Represents variant-specific tasks
     |    |--- NewIncrementalTask
     |    |--- NonIncrementalTask
     |--- NonIncrementalGlobalTask
     |--- AndroidLintGlobalTask

That means we currently don't have a class to represent global
(non-variant-specific) tasks.

This leads several inconsistencies and confusing setup (as we will see
in the contents of this commit).

To address that, this commit will introduce AndroidGlobalTask, so the
hierarchy is clearer:
     BaseTask
     |--- AndroidVariantTask: Represents variant-specific tasks
     |    |--- NewIncrementalTask
     |    |--- NonIncrementalTask
     |--- AndroidGlobalTask (NEW): Represents global tasks
          |--- NonIncrementalGlobalTask
          |--- AndroidLintGlobalTask

Also update the `TaskCreationAction`s to reflect the new hierarchy.

Bug: 300617088 (In progress)
Test: Updated AnalyticsUtilTest, TaskMethodModifiersAndAnnotationsTest,
      BuildAnalyzerTest
Change-Id: I0c969d1a82359c73143101e57391dabd56948f64