Switch to the analysis API to more accurately detect non-returning calls
Until now we've had a hardcoded list of methods known not to return,
such as the Kotlin standard library methods `error` and `TODO`.  This CL
updates it to use the analysis API instead, such that it can look for
methods that use the Nothing return type. This will pick up for example
`exitProcess` and `numberFormatError` and additional methods in the
user's own code.

Bug: 135714717
Test: VersionChecksTest.testExceptionsAndErrorsAsExitPoints (updated)
Change-Id: I4bf3eb944e1c65251237efa0a7b84b74d9598442