Stop assuming no main manifest means no overlays
When processing test manifests, code was assuming that if the main
manifest file was absent, it meant that no merging was necessary.
However, it's legal to have a manifest file just in the overlays
(like build-type/product-flavor).
To complicate things further, generated manifests added through
the variant API are overlays to the main manifest.
In case no main manifest file exists in the source folders, elect
the highest priorty overlay (if there are any) as the main
manifest and proceed as usual

Test: Added integration tests
Bug: 341124476
Change-Id: Ida4ecc4db2f0e1d2c8a8c464d795a7713b6f36d3