Do not add static directories twice
LayeredSourceDirectoriesImpl variantSources and variantStaticSources
are holders of internal data structures. When adding static directory,
the directory was added to both holders which eventually created
a duplicate situation where the directory was present twice (generating
duplicate merging errors for instance). Also when merging directories
in the model builder, use a Set<> to remove all duplicates for the IDE.

Test: Added a regression test
Bug: 342428022

Change-Id: Ieb74db2175e4d64a6a011bc864d6d043c1917caa