Fix and move CredentialDependency issue
Move issue id "CredentialDependency" to a new Detector class. Now
reports the incident on the <application> tag of the manifest to make
it clear which app module is problematic, and to detect the problem even
when we can't see any Gradle statement for the dependency. We now also
skip reporting the problem if the app is a Wear OS module (based on the
manifest).

Bug: 339363983
Test: unit tests updated and checked manually in Studio
Change-Id: I57e7706fb35dfd0767a369a4f3a47f12906cf63c