Fix lint for composite builds
I722a2d907d broke lint for some composite builds. The reason for the
breakage was because of the assumption that the buildTreePath is always
equal to (buildId + projectPath).replace("::", ":"), which isn't the
case for composite builds.

This change reverts parts of I722a2d907d to avoid this string
manipulation.

Fixes: 327670497
Test: CompositeBuildTest
Change-Id: I039fde226805d2593c11306caae31e310fb5a68c