Add AvdBuilder for editing AVDs
Currently, we have AvdInfo that holds most of the configuration state of
an AVD, but to create or edit an AVD, we need to invoke a method with 11
arguments.

Several of these arguments are low-level string maps; the caller must
handle details like representing true/false as "yes" and "no" and
parsing storage specs like "512M".

We also don't have a way to easily go from the AvdInfo object to the
parameters that need to be passed to createAvd. Doing it naively is
error-prone; replacing the device, skin, or system image without
clearing its implied properties will result in the previous values
mixing with or overriding the new ones. Currently, this logic resides in
UI code.

Bug: 339875083
Test: AvdBuilderTest
Change-Id: Iee5298db796764f6efe95a2ab5fea316edea31bd