Use target_compatible_with instead of tags to specify platform compatibility.
target_compatible_with skips targets that have a transitive dep on incompatible targets. This lets a lot of previous no_* tags be omitted since they depend on incompatible targets.

Some no_* tags are kept for now because they are used to disable testing. They will be replaced in another change with the appropriate test opt-out tag.

Bug: N/A
Change-Id: I25a0238b52e2862d91c4198da13b17f96e41c565
Test: presubmit