Support built-in Kotlin for compose projects
Compose projects using Kotlin 2.0.0+ still aren't compatible with
built-in Kotlin because the Compose Compiler Gradle plugin is not
compatible wiht built-in Kotlin. After this change is merged, I'll file
a bug against Jetbrains for this.

This change adds Kotlin 1.9.23 in order to remove the
"suppressKotlinVersionCompatibilityCheck" flags currently used in
some AGP tests.

Bug: 341765853
Test: ComposeHelloWorldTest
Change-Id: I4d9e75fd072257a397073fbc1ac642f693815d1b