Handle more binary resolution cases in ApiDetector
In addition to `PsiCompiledElement`, for facade class in particular,
`KtLightClassForDecompiledDeclaration` (for both K1 and K2) can be used
to represent (literally decompiled) classes from binary dependencies.

Instead of deserialized descriptor in K1, UAST in K2 IDE(!) has "source"
of even stdlib. But, for the consistency, it "fake"s resolution result
for those from deserialized declaration. In doing so, it uses the facade
class rather than multi-file class part.

Bug: 350744703
Bug: 351955861
Test: AndroidLintTest#testApiCheckKotlinExtension w/
  -Didea.kotlin.plugin.use.k2=true
Test: ApiDetectorTest w/ -Dlint.use.fir.uast=true
Change-Id: Icc00e1fa7dc2e0bcb5be4982b8c7dc9d053b6cc2