[SDK Index] Add issue description as note
... for messages from developers.

Recent versions of the SDK Index snapshot contain a description for the
critical issues (messages from developers). Now we can display that
infromation in Android Studio. This change adds that description in the
genererateBlockingCriticalMessage and generateCriticalMessage methods so
users can see those descriptions without needing to navigate into the
SDK Index.

This change applies commit 3cf50e16beb9764b42be889381f073b4a07b41b2
again but adds a flag to control if the notes are shown or not.

Bug: 336568007
Test: GooglePlaySdkIndexTest
Change-Id: Ibbe0a066c76f0c6064c3590df755379cf9bd32c6