only process static manifest file during model building
the model builder is processing manifest file to look for
instant app tags. this can only be done for static files but
not generated ones that depends on a Task execution.

this mean that users will not be able to set the instant app
flag in a generated manifest file. It must be set in a static one.
Eventually, we should stop this manifest file parsing during
model building but instead use a DSL flag, see b/160970116.

Test: Added a regression test.
Bug: 343076001
Change-Id: I47a53fbda11a7bfcbedd1f66ce6e6e3001cb13a2