Prevent resuming already resumed continuation
There is a rare bug where in `ChannelReadOrWriteHandler` we
resume a `continuation` that was previously resumed.
This is possibly happening, because the channel is closed right
after the completion handler has been called. Adding a check for
`completion.isActive` should prevent us trying to resume completion
twice.

This change also refactors `ChannelReadOrWriteHandler` to
confine the `resume` and `resumeWithException` calls
to the completion handler's `complete` and `fail` methods
as well `completionHandlerCompleted` method directly called
by them.

Test: existing
Bug: 338008891
Change-Id: Icda70549bcd4286cbd6cc0696dd502bf100350ae