[SDK Index] Show warning only in 3rd party libraries
We currently display a message that says "These versions have not been
reviewed by Google Play. They could contain vulnerabilities or policy
violations. Carefully evaluate any third-party SDKs before integrating
them into your app" whenever we recommend a version in the SDK index
issues, regardless of the author of the library.

This change checks if the library is a third party before adding that
part in the issue message. The check is done using the information in
the SDK Index snapshot.

This change is based on commit a2f506918ccf87e25a891b61b3b6f21fe8fec25a
and 12dfed2341fa95b639a99a9af34adc140f7d827e so gmaven is not used.

Bug: 338654964
Test: GooglePlaySdkIndexTest, GradleDetectorTest
Change-Id: I13cf24bebe10ed3b1812ab8573b90349690b5e2f