Add built-in KAPT support for all components
This change adds a new plugin, "com.android.legacy-kapt".

If that plugin is applied (or if we're in a kotlin-enabled testFixtures
component or a screenshotTest component and the existing kotlin-kapt
plugin is applied), AGP will register and wire KAPT tasks using the
KotlinBaseApiPlugin.

The new plugin has "legacy" in its name to indicate that users should
migrate to KSP instead of using it.

This change is based on patchset #45 of I35d20ba04d

Bug: 341765853
Test: BuiltInKaptTest, etc.
Change-Id: I627ccbaf0cdd1387a02af850b8f406e103e463c4