Support annotation references in XML files too
Classes can be referenced from XML files (such as layout files, and
manifests, and so on). This CL updates the annotation machinery such
that lint checks can easily handle class lookups related to annotations.

It updates two detectors for this:

(1) The @DiscouragedApi detector (which was the motivation for this
    fix), and
(2) The API detector, such that an @RequiresApi annotated view
    referenced from XML starts creating warnings.

(Also fixed some warnings in nearby code.)

Test: Unit test included
Bug: None
Change-Id: I6d648778a085a8d39674ad7920924585d76f80d2