Fixes for KTS handling and namespace handling
First, this CL fixes the handling of namespace attributes.  In
particular, in lint, the way to look up the namespace for the project is
using the "package" attribute on projects, but this was returning null
in a few scenarios. One such scenario is now captured by the new
ManifestDetectorTest (testDuplicateMissingPackage) which until this CL
included "null" in the error message for the package.

(1) On the IDE side. It had been stubbed out as "null" in
    LintModelFactory with the comment that it's "not in the old builder
    model", but it's there now, so hook that up.

(2) In lint unit test projects with mocking of the Gradle model.

(3) The code to resolve manifest attributes was only consulting the
    package attribute in the same file, stating that it's required to
    exist -- but that stopped being the case a few years ago, and our
    upgrade assistant even proactively removes these from manifest files
    in favor of the namespace declaration in build.gradle.kts.
    Therefore, update this and update the callers of this method to
    include enough context to also look up the project namespace.

This second discovery led to a bunch of related fixes:

(4) KTS files were not getting the same treatment as regular Gradle
    files in the testing infrastructure.  Groovy-based files use a light
    interpreter to create a corresponding lint model (and ensures that
    the whole project starts looking for tests in source sets (such as
    src/main/java and src/main/res) rather than just src/. This CL hooks
    this up, and updates a number of KTS tests accordingly.

(5) The Gradle model mocker was updated to handle various KTS
    constructs.

Test: Existing (many updated)
Bug: Not filed
Change-Id: I38bbd679ecf822dffb7e9912ceb709155932812a