Create dedicated class to describe targeted shrinking rules
Targeted shrinking rules were introduced in AGP 3.6 (bug 135672715).
The logic for handling these rules is a bit complex, and currently this
knowledge is scattered in various places (ExtractProGuardRulesTransform,
AarTransform, and FilterShrinkerRulesTransform), making it even harder
to discover.

To address that, this change creates a dedicated class called
TargetedShrinkRules to document these rules as clearly as possible.

In an upcoming change, we will migrate existing code to use the new APIs
in TargetedShrinkRules.

This will make it easier to support the new R8 keep annotations, which
introduce additional complexity.

No-Typo-Check: Linter wants to change "upto" to "up to", but in this
case "upto" is part of a valid file path.

Bug: 340781050
Test: New TargetedShrinkRulesTest, TargetedShrinkRulesIntegrationTest
Change-Id: I320b4f4034d05556fede4aed5f00c01a889efcb2