Add MapBinding for AVD config I/O
AVD config is stored in a text file as a list of key/value strings.
Most of our code maintains this low-level representation; methods in
AvdManager and AvdInfo expose it as a Map<String, String>. It is mostly
left up to the UI layer to do the appropriate conversions to
higher-level types.

Add the ValueConverter interface to define the conversion to and from
String. Add MapBinding to define the conversion from a property to a
map entry. Add extension methods for conveniently setting up the
binding.

Bug: 339875083
Test: MapBindingTest
Change-Id: I17c50cde762034246d5c547cc10d26f80dcf00b0