Suppress Exceptions if the APK File is Abandoned
The existing implementation is severely flawed.

It copies the APK into a temporary file. Then it registers a listener
for file system changes and if it detects the original file changes,
it refreshes the view by deleting the tmp file and creating a new one.

There are 2 issues with this system:
1. It doesn't check if the contents actually changes. So it triggers
even when no changes are made (It's not clear to me why the
VirtualFileManager.VFS_CHANGES is triggered when the file contents are
not actually changed).
2. When the file-changed event is handled, we delete the file and create
a new one. But there's a lof of underlying objects that are still
attached to the old temp file and they try to reload it.

I fixed issue #1 but issue #2 is still possible if the file does
actually change. I could not fine a way to handle this cleanly. As in,
prevent the code from trying to access the closed/deleted file. Instead,
handle the exception thrown and if the file was closed or deleted, I
ignore the exception and let the code fail silently at the call site.

Bug: 348406593
Test: Added
Change-Id: I4e4be5be4fe429e103bfc0ef75fe9616480871bb