refactor deviceTests APIs to return a Map
The incubating API was returning a List<> which made it
difficult to disambiguate the instances outside of using the
component name which is built from the product flavor/build type
combination.

With the new API, one can obtain the DeviceTest and DeviceTestBuilder
providing a simple name like "AndroidTest". So far, there is only
one type of device tests but this might evolve in the future.

Test: Existing
Bug: N/A
Change-Id: Ibfdf9653600df1dbc3e32fa770db99320df0a9c9