Catch additional obsolete SDK_INT checks
The SDK_INT checks in the source code are checked to see if they are
always true, or always false, based on the current minSdkVersion.

However, there are cases where we have additional clues in the source
code (such as @RequiresApi, and surrounding SDK_INT checks) where we can
also conclude that an SDK_INT check is unnecessary. This CL implements
and tests that.

Bug: Not filed
Test: Unit tests included
Change-Id: I6bc83d42eaeeb34fa612ce3036ccafa8b8dff9db