[SDK Index] Use SDK index as another source of versions
... when suggesting changes in dependency versions.

The SDK Index snapshots contain a flag telling if a given version is the
latest one for a particular library. Android Studio has some quick fixes
to suggest users to change a library version by looking into the cached
libraries, gmaven and other sources. This change adds the snapshot
latest version as a possible suggestion if it is not preview.

Bug: 336826464
Test: GradleDetectorTest, GooglePlaySdkIndexTest
Change-Id: I1e40f1195d01251553b9dc3a5335872676a8fc6d