[SDK Index] Vulnerability issues should be warnings
... by GooglePlaySdkIndex#hasLibraryErrorOrWarning.

Vulnerability issues support was added to Android Studio but versions
that contain those issues are not currently considered to be errors or
warnings when checking if a library has them. This could result in users
gettng quick fixes that will replace library version with one that has a
vulnerability issue. This change adds vulnerability issues when checking
for errors and warnings.

Bug: 361415883
Test: GooglePlaySdkIndexTest
Change-Id: Ifc5f27ae10b98bc309e0afd9f78cf41f648f7fa8