ResourceNameKeyedMap: remove trove4j usage
Only trove4j and fastutil seem to support maps with a custom hash
function. However, trove4j is deprecated (b/357843093) and fastutil
is a 20 MB dependency (not quite worth having in sdk-common).
So, instead I'm wrapping keys in a wrapper class.

Historical note: the previous version of this code comes from
commit 3a1ff530ab / Change I1e1be10b1, which was extra careful to
avoid String allocations during get() calls. The current change
regresses slightly by allocating a KeyWrapper() instance on each
get() call. However, a KeyWrapper is very small, the instances are
short-lived, and I expect the impact on performance to be negligible.

Note: the keySet() method is omitted for simplicity; there were no
clients using it aside from the unit test.

Bug: 357843093
Test: ResourceNameKeyedMapTest (updated)
Change-Id: I7a86bba4c52b36db6a70ba2857ebcc6966456bee