[dex] Sort dex files for bundling
Dex classes files are not added in a consistent order compared to APKs
and art profiles that pre-sort using DexFileComparator. This is not the
case for writing dex to bundles.

This has been causing issues with the profgen tool which expects the dex files used to
create the input baseline profile to be identical to the dex in an apk
derived from a bundle. This is not the case when core library desugaring
is enabled and adds the desugar_lib_dex to a differing position.

Bug: 346268213
Test: CoreLibraryDesugarTest
Change-Id: I669558c636a7144f24f575c71ad2f2caed21a7f3