Always report permission check missing error
PermissionDetector previously would not report an incident for a missing
permission check if it reported a (conditional) incident for a missing
manifest permission. However, this could result in neither incident
being reported, as the conditional incident could get filtered out.

This change makes it so both incidents are reported (as conditional
incidents). Also, moved the targetSdk check (for missing permission
check incidents) to filterIncident (unless we are in a global analysis).

Bug: 204897515
Test: added
Change-Id: I200fbaa23eff1d5930b6f0230291c04060e8398d