Avoid creating same AA modules
Before Ic41ab127fb840623a90272ef967f61115a3f5c39, Project and its
dependencies are topologically sorted according to their "name"s.
It may discard Projects with the same name (due to the same last part of
the directories, like `out`), but had a side effect of making the visit
and AA module creation happen only once.

AGP Lint tasks put such cache-like `out` to the classpath of the main
module, so it's not really necessary to build/maintain/wire the library
module module for that. We determine that as a project without source
files to analyze. Filtering them out confirms to work with benchmarks.

Bug: 371220733
Test: bazel test \
  --override_repository=yourkit_controller=~/Downloads/yourkit-controller \
  //prebuilts/studio/buildbenchmarks:Base100_Java8_lintAllModules_profiled
Change-Id: Iaac5117d6aa427771cfe0226d0184b0aaae464f8