Ignore regular android views if xr views are present
If there are xr panels, we want to ignore regular android views.
In practice, there are no regular views in an xr app,
but the regular view APIs still return the main panel views as
a regular view. So getting views both using xr and regular view
APIs would duplicate the main panel.

Test: NA
Bug: NA
Change-Id: If4636538a3e2ac745136c8913bda3da4f5e7ceb2