Provide provenance on merge conflicts
When a java resource merge conflict happens, we now
provide the provenance for each of the artifact that
conflicted.

This requires bringing the name of the dependencies
or component into the task action and into the merger.
The merger already had support for names inside the
IncrementalFileMergerInput class so we reuse that
(making sure that the error uses it).

On the task side, we keep track of the provenance
by using ArtifactCollection as long as we can
instead of just using FileCollection in the inputs.

Fixes: 373462080
Test: existing tests plus new ones specifically for conflits
Change-Id: I88d82df521f3f15b609a8a2d8ada5eb6a749359e