Make Gradle Executor more flexible.
Cut ties with GradleTestProject mostly via GradleTestInfo
(formerly known as GradleTestRule).

GradleTestRule used to extend TestRule but that's really
not necessary. The goal is to instead provide some info
regardless of the object (rule in this case, as we're looking
at introducing a new rule).

This also encapsulate some options that are common to the rule
and the executor in an attempt to reduce code duplication.
It's not super successful due to our usage of the builder
pattern which makes using delegates very difficult. We should
probably revisit how all these parameters are configured.
At least for now we don't have to end up with more and more
objects passed around.

Bug: N/A
Test: N/A -- this is test fixtures so all tests validate changes.
Change-Id: I7bccf5f101a651ef1602c1475cc009ad366622f7