Make TemporaryProjectModification usable by other classes.
Introduce FileProvider as this is the only thing we need
from GradleTestProject. Any implementation can now benefit
from this.

Also introduce a delegate for subproject that can share the
same event map. The top level instance can be created
without a FileProvider to represent a Build rather than
a subproject.

Bug: N/A
Test: This is test fixtures
Change-Id: I055ea38089e94b99cd3c96d59af3350e616bd110