Ignore android.jar when computing desugaring graph for dexing
In dexing transforms, for the desugaring graph to be relocatable, we
require that all nodes (files) are located within the project directory.

However, in some cases, D8 may report dependencies between the project's
classes and android.jar, which is located outside the project directory,
and the build would fail.

To fix this issue, we will now ignore those dependencies instead. It is
safe to ignore because:
  - `bootClasspath` is a non-incremental transform input, so any changes
  to it will cause the transform to run non-incrementally (i.e., we
  don't need to track changes to `bootClasspath` and files that are
  impacted by those changes).
  - `bootClasspath` is self-contained, it doesn't depend on anything
  else, so we don't need to track its dependencies.

Bug: 362339872
Test: Added DesugarGraphTest
Change-Id: I677122b893a9015e95cb14a67483ee5ef2070c4a