Test android.nonFinalResIds=true/false in ResourceShrinkerTest
`android.nonFinalResIds` is enabled by default, but ResourceShrinkerTest
(and some other tests) were disabling it to verify certain behavior
(see commit 3347482).

The issue is that, in an upcoming change, we will introduce R8's
optimized shrinking feature, which requires `android.nonFinalResIds` to
be enabled.

Therefore, we'll now run ResourceShrinkerTest with both
`android.nonFinalResIds` enabled and disabled.

Bug: 326033134
Test: Updated ResourceShrinkerTest
Change-Id: I90146a10e817efc8dd221522f52c1a45af5e9dbe