Adapt to KT UAST 243
* No more fake resolution results for primitive operators
https://youtrack.jetbrains.com/issue/KTIJ-30636
https://github.com/JetBrains/intellij-community/commit/2d1b15d7b05635b0144612608222e07cfac62977

The real headache was that all the recent improvements on
TypedefDetector were using that fake resolution results:
https://googleplex-android-review.git.corp.google.com/q/owner:jsjeon@google.com+TypedefDetector

In this commit, I simply reinvented the wheel: same logic in AA

* Resolution for simple property access
https://youtrack.jetbrains.com/issue/KTIJ-31140 from b/308011060#comment43
https://github.com/JetBrains/intellij-community/commit/69ad4effce4a9a22231480eea2a6802a4b73a47f

This bothers several detectors to raise duplicate reports:
originally from simple name reference's accessor call visit,
now the 2nd visit for resolved operator.
Since all the related places in Lint detectors/API assumed that this
binary operator resolution is for "overloaded" operator, I added utils
to explicitly find that: operator except for simple property access.

* Error type for instantiation of local class
https://youtrack.jetbrains.com/issue/KTIJ-17870
https://github.com/JetBrains/intellij-community/commit/db5f01f6a45d54605f66fecb00e7bacccce590bf

But, if that class is delibrately error type (or not exist),
one detector checks if the type is UastErrorType. For some case,
it can be a PsiClassReferenceType with error/NonExistentClass.
I added a util to consider that kind of error type too.

Bug: 367868888
Test: bazel test //tools/base/lint/libs/lint-tests:tests_k2.test
Test: bazel test //tools/base/lint/libs/lint-tests:tests.test
Change-Id: I9db132597b2d7129614fcb0a0b77b089c72992f8