Correctly calculate package id from resource identifier
Package ids in the proto based interpretation of the resource table is an unsigned byte.

In the current package mapping calculation we would get -128 for 128 since the package id is the uppermost byte, which we would shr down 24

I have added regresion testing of this in the R8 repo (and landing the same fix).

The regression test can be found here:
https://r8-review.git.corp.google.com/c/r8/+/97980

Cherry picking the fix to studio for consistency (we are moving over to calling this through R8)

Test: added in r8 repo, see above
Bug: 378470047
Change-Id: Ibe403fc8bfe71f1349d5eb1ab5da97afdf69fbec