374749030: Allow `--disable` for any check with `--XallowBaselineSuppress`
You can now disable issues that use suppressAnnotations=listOf(..) with
a special invocation flag, --Xdisable <id>. This is like --disable, but
also works (and doesn't warn) for suppression annotation issues.

Also fixes a couple of leftover cases where lint should allow
suppressing issues of this type if the suppression id is the same as the
issue id.

As part of the unit test for this I wanted a lint check which actually
reports an error, so added a few more stubs into the lint api stubs
so allow using a small subset of the source scanner API.

Fixes: 374749030
Test: Unit test included
Change-Id: I4c64c1940b738d34bac6775318758b14aa39bda2