Consider uses-sdk's featureFlag attribute when validating merged manifest
After manifest merger is complete, we validate that there is only
a single uses-sdk element in merged manifest file. This check is now
updated to also consider the featureFlag attribute on uses-sdk element.
Now we use element's id which include featureFlag attribute value to
detect duplication. Once duplication is detected, we find the element's
node and attribute records to find the source manifest file.

Bug: 378158206
Test: ManifestMerger2Test
Change-Id: I5b674a78ffaa62acce3c3fa70eb185b9d6491189