Handle AC restart for API 30+ via recreate()
Changes to API 35+ prevent update-appinfo from restarting the activity
in the case where there is no actual change to ApplicationInfo. This
prevents the use of am update-appinfo for restarting the activity in
Apply Changes.

This change modifies overlay swap (which occurs for devices with API
30+) to use the agent to manually restart the activity after the swap
operation and overlay update complete.

This also requires updating the instrumentation for IWI-on-swap, as the
update-appinfo callback can no longer be instrumented to update resource
loaders; resource loaders are now updated by the agent directly prior to
activity restart.

Fixes: 361737406
Test: Existing E2E coverage
Change-Id: Ife843643f0b7490db0418d3161999b95574e2c0d