Handle standalone artifacts in lint's doc generator
This CL updates lint's issue documentation generator such
that it handles the new standalone artifacts released by
AndroidX. As of recent versions, all the compose lint checks
are released both as part of the compose library AARs, as
well as standalone. This initially showed up as duplicate
issues and libraries. This CL updates this such that the
issues are deduplicated, and instead, the artifact page
for these libraries also points to the corresponding aliases.

Instead of just basing this on Gradle dependency string
manipulation, this CL looks at the lint jar binaries and
identifies duplicates. This makes it also handle some cases
where a lint check was duplicated across multiple artifacts,
such as the various androidx.wear checks.

There still are some cases of duplicate checks (where the
lint jars aren't actually identical). One such scenario is
reported b/379118860. But I discovered that the index page
linking to the duplicate issues was actually all pointing
to the default page, so this CL also fixes that.

To see the net effect of these changes on the official
documentation, see
https://github.com/googlesamples/android-custom-lint-rules/commit/8a265e3642e6b0c2802841720b9ea1782c7d2304

Test: Existing LintIssueDocGeneratorTest, and manual
    inspection of diffs in update to
    https://googlesamples.github.io/android-custom-lint-rules/checks/index.md.html
Bug: Not filed
Change-Id: I0899f7d1895b65ae5095e75300b0fa46f1ee6812