GradleRule: more consistent API.
The API to get a GradleBuild from the rule was not consistent.
This commit changes it so that directly getting a build or first
configuring it uses the same names. It also moves away from
the from() method as this was already used by the static methods
on GradleRule and could make things more complicated. The API
is now like this:

// to get a rule object
(fun) GradleRule.from { }
(fun) GradleRule.configure().from { }

// to get a build object from the rule
(val) rule.build
(fun) rule.build { }
(val) rule.configure().build
(fun) rule.configure().build { }

Bug: N/A
Test: existing
Change-Id: I3e659e24cfbde61113e7d488583e0e26860d3b7c