GradleRule: refactor Android support classes.
Each Android plugin requires 3 different sets of
classes and interfaces:
- the definition. This drives the DSL to describe
  the project
- the AndroidProject. This how test code interact with
  the created project
- Reversible project: this allows making modifications
  and reverting them.

Previously we had one files for each of these types,
that contained the interfaces and classes for _all_
android plugin types. This was not manageable as we
need to add support for more types (test, aiPack, fused
library, etc...)

This refactor move all types related to a plugin type
into a single file. This also introduces per-plugin
type interfaces for GradleProject.

Finally, this refactors code in the GradleBuild
implementations to reduce duplication

Bug: N/A
Test: N/A
Change-Id: I93c3f6ed5961f9403948af7c4097f73e01005356