Fix issue MergeJavaResourceTask failing when file is changed
A failure is occurring in the MergeJavaResourceTask when a file is
changed (incremental build). There are unchanged files incorrectly being
processed in the incremental build when an artifact contains multiple
files. For instance, for "androidx.emoji2_emoji2.version", there is a
"classes.jar" and a "repackaged.jar" file, but along with the recent
change to add provenance, `inputsForFile` in the IncrementalFileMerger
is matching with two entries (since the identifier is just the artifact
owner's name rather than the absolute path). Then, the merger tries to
update the file even though it was not changed, leading to the error
message "Unknown file ..." in LazyIncrementalFileMergerInput.

This adds the file name to the input so that the distinction can be
made and it is not mistaken as an updated file.

Bug: 377366954
Test: updated
Change-Id: I114776b76f487249e196fc73647e6ab6f85a2539
(cherry picked from commit ae5f9ef4abe0d261ab4fb7ed86d3e5ea81847873)