GradleRule: migrate ReferenceModelComparator
In order to support having 2 rules in the same
test, this change adds the ability to customize
the root folder of the generated project, and
the name of the build (which is written in settings)
so that the reference and delta projects are written
in different location but have the same name (as
the name is used to infer the groupId of projects).

A big change in some of the snpashot is that the
original tests all used root project for android
app and lib, which is not possible in the new
fixture. This has an impact on a few snapshots.

Bug: n/a
Test: migrating tests to new fixtures
Change-Id: I3276d480e752aa6f32df4301097245f52980e594