threading-agent: expand heuristic for lambdas
Background: the Kotlin K2 compiler uses invokedynamic by default for
lambdas, which involves generating a synthetic method for the lambda
body. Surprisingly(?), the generated method is not marked with
ACC_SYNTHETIC. Instead it looks like this:

  private static final kotlin.Unit main$lambda$0();
    descriptor: ()Lkotlin/Unit;
    flags: (0x001a) ACC_PRIVATE, ACC_STATIC, ACC_FINAL

In order to preserve the current threading-agent behavior (in which
class-level threading annotation do *not* apply to lambdas), we need
to manually exclude methods that look like Kotlin lambdas.

Without this change, tests such as LayoutInspectorViewIntegrationTest
fail due to false-positive threading violations.

Test: LayoutInspectorViewIntegrationTest (for example)
Bug: this change unblocks b/374993754
Change-Id: Ie9071f8ace250ce40a26fbf32cb60ed4c161bf4c